gather all tests related to scaling, translating and rotating in one place
