+++ src/commonTest/kotlin/com/jillesvangurp/geogeometry/GeoGeometryTest.kt
@@ -93,39 +93,6 @@ class GeoGeometryTest {
     }
 
     @Test
-    fun shouldRotate() {
-        val anchor = bergstr16Berlin
-        val point = oranienburgerTor
-        val d = GeoGeometry.distance(anchor, point)
-        val points = (0..240).step(10).map {
-            GeoGeometry.rotateAround(anchor, point, it.toDouble())
-        }
-            .also {
-                // all points should be the same distance
-                it.forEach { (GeoGeometry.distance(bergstr16Berlin, it) - d).absoluteValue shouldBeLessThan 0.1 }
-                it.size shouldBe 25
-                it.distinct().size shouldBe 25
-                it.contains(bergstr16Berlin) shouldBe false
-            }
-
-        val features = (points + point + anchor).map {
-            Geometry.Point(coordinates = it)
-        }.map {
-            it.asFeature()
-        }
-        println(FeatureCollection(features))
-    }
-
-    @Test
-    fun rotateByZeroDegreesShouldBeSamePoint() {
-        val anchor = bergstr16Berlin
-        val point = oranienburgerTor
-        GeoGeometry.distance(point, GeoGeometry.rotateAround(anchor, point, 0.0)) shouldBeLessThan 1.0
-        (GeoGeometry.distance(point, GeoGeometry.rotateAround(anchor, point, 180.0)) - 2 * GeoGeometry.distance(anchor,
-            point)).absoluteValue shouldBeLessThan 1.0
-    }
-
-    @Test
     fun headingFromTwoPoints() {
         GeoGeometry.headingFromTwoPoints(
             doubleArrayOf(13.0, 52.0),
