+++ versions.properties
@@ -11,7 +11,8 @@ version.junit.jupiter=5.10.1
 
 version.kotest=5.8.0
 
-version.kotlin=1.9.21
+version.kotlin=1.9.22
+## # available=1.9.22
 ## # available=2.0.0-Beta1
 ## # available=2.0.0-Beta2
